+++ src/Data/Response/SdkResponse.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Vin\ShopwareSdk\Data\Response;
+
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
+
+class SdkResponse extends Response implements ResponseInterface
+{
+}