+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to `shopware-php-sdk` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+### 1.2.0
+- Add ActionButtonResponse classes that can be return directly when receive an action button request
+- Add RegistrationResponse that can be return directly when receive an app register request
+- Update Docs for App
+
 ### 1.1.0
 - Add plugin custom Definition at runtime
 - Update all latest entities from SW 6.4.2.1